+++ src/basic/types/product.type.ts
@@ -1,0 +1,6 @@
+export type Product = {
+  id: string;
+  name: string;
+  originalPrice: number;
+  quantity: number;
+};
